core: convert project to PCRE2
Convert project to PCRE2 as PCRE is EOL and won't receive any security
updates anymore.

PCRE2 changed some API and concept. Mainly there isn't the concept of
study anymore, replaced by match_context concept and match_data is used
instead of ovector to handle results. Because of this the scratcher is
not needed anymore and is replaced by a simple function to setup the max
ovector size on end module init.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>